YQL-19747 Extract ANTLR4 module
