+++ yql/essentials/sql/v1/complete/antlr4/c3i.h
@@ -1,0 +1,40 @@
+#pragma once
+
+#include "defs.h"
+
+#include <util/generic/fwd.h>
+#include <util/generic/string.h>
+#include <util/generic/vector.h>
+
+#include <unordered_set>
+
+namespace NSQLComplete {
+
+    struct TSuggestedToken {
+        TTokenId Number;
+    };
+
+    struct TMatchedRule {
+        TRuleId Index;
+    };
+
+    struct TC3Candidates {
+        TVector<TSuggestedToken> Tokens;
+        TVector<TMatchedRule> Rules;
+    };
+
+    class IC3Engine {
+    public:
+        using TPtr = THolder<IC3Engine>;
+
+        // std::unordered_set is used to prevent copying into c3 core
+        struct TConfig {
+            std::unordered_set<TTokenId> IgnoredTokens;
+            std::unordered_set<TRuleId> PreferredRules;
+        };
+
+        virtual TC3Candidates Complete(TStringBuf queryPrefix) = 0;
+        virtual ~IC3Engine() = default;
+    };
+
+} // namespace NSQLComplete
