+++ yql/essentials/sql/v1/complete/antlr4/c3t.h
@@ -1,47 +1,21 @@
 #pragma once
 
-#include "sql_antlr4.h"
-#include "string_util.h"
+#include "c3i.h"
+
+#include <yql/essentials/sql/v1/complete/text/word.h>
 
 #include <contrib/libs/antlr4_cpp_runtime/src/ANTLRInputStream.h>
 #include <contrib/libs/antlr4_cpp_runtime/src/BufferedTokenStream.h>
 #include <contrib/libs/antlr4_cpp_runtime/src/Vocabulary.h>
 #include <contrib/libs/antlr4-c3/src/CodeCompletionCore.hpp>
 
+#include <util/generic/fwd.h>
 #include <util/generic/string.h>
 #include <util/generic/vector.h>
 
 #include <unordered_set>
 
 namespace NSQLComplete {
-
-    struct TSuggestedToken {
-        TTokenId Number;
-    };
-
-    struct TMatchedRule {
-        TRuleId Index;
-    };
-
-    struct TC3Candidates {
-        TVector<TSuggestedToken> Tokens;
-        TVector<TMatchedRule> Rules;
-    };
-
-    class IC3Engine {
-    public:
-        using TPtr = THolder<IC3Engine>;
-
-        // std::unordered_set is used to prevent copying into c3 core
-        struct TConfig {
-            std::unordered_set<TTokenId> IgnoredTokens;
-            std::unordered_set<TRuleId> PreferredRules;
-        };
-
-        virtual TC3Candidates Complete(TStringBuf queryPrefix) = 0;
-        virtual const antlr4::dfa::Vocabulary& GetVocabulary() const = 0;
-        virtual ~IC3Engine() = default;
-    };
 
     template <class Lexer, class Parser>
     struct TAntlrGrammar {
@@ -68,29 +42,24 @@ namespace NSQLComplete {
             CompletionCore.preferredRules = std::move(config.PreferredRules);
         }
 
-        TC3Candidates Complete(TStringBuf queryPrefix) override {
-            Assign(queryPrefix);
-            const auto caretTokenIndex = CaretTokenIndex(queryPrefix);
+        TC3Candidates Complete(TStringBuf prefix) override {
+            Assign(prefix);
+            const auto caretTokenIndex = CaretTokenIndex(prefix);
             auto candidates = CompletionCore.collectCandidates(caretTokenIndex);
             return Converted(std::move(candidates));
         }
 
-        const antlr4::dfa::Vocabulary& GetVocabulary() const override {
-            return Lexer.getVocabulary();
-        }
-
     private:
-        void Assign(TStringBuf queryPrefix) {
-            Chars.load(queryPrefix.Data(), queryPrefix.Size(), /* lenient = */ false);
+        void Assign(TStringBuf prefix) {
+            Chars.load(prefix.Data(), prefix.Size(), /* lenient = */ false);
             Lexer.reset();
             Tokens.setTokenSource(&Lexer);
-
             Tokens.fill();
         }
 
-        size_t CaretTokenIndex(TStringBuf queryPrefix) {
+        size_t CaretTokenIndex(TStringBuf prefix) {
             const auto tokensCount = Tokens.size();
-            if (2 <= tokensCount && !LastWord(queryPrefix).Empty()) {
+            if (2 <= tokensCount && !LastWord(prefix).Empty()) {
                 return tokensCount - 2;
             }
             return tokensCount - 1;
