+++ yql/essentials/sql/v1/complete/antlr4/ya.make
@@ -1,0 +1,13 @@
+LIBRARY()
+
+SRCS(
+
+)
+
+PEERDIR(
+    contrib/libs/antlr4_cpp_runtime
+    contrib/libs/antlr4-c3
+    yql/essentials/sql/v1/complete/text
+)
+
+END()
