+++ yql/essentials/sql/v1/complete/sql_antlr4.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <yql/essentials/sql/v1/complete/antlr4/defs.h>
+
 #include <contrib/libs/antlr4_cpp_runtime/src/Token.h>
 #include <contrib/libs/antlr4_cpp_runtime/src/Vocabulary.h>
 
@@ -7,12 +9,7 @@
 
 #include <unordered_set>
 
-namespace NSQLComplete {
-
-    using TTokenId = size_t;
-    using TRuleId = size_t;
-
-    constexpr TTokenId TOKEN_EOF = antlr4::Token::EOF;
+namespace NSQLComplete {    
 
     class ISqlGrammar {
     public:
