+++ yql/essentials/sql/v1/complete/sql_complete.cpp
@@ -1,8 +1,8 @@
 #include "sql_complete.h"
 
 #include "sql_context.h"
-#include "string_util.h"
 
+#include <yql/essentials/sql/v1/complete/text/word.h>
 #include <yql/essentials/sql/v1/complete/name/static/name_service.h>
 
 // FIXME(YQL-19747): unwanted dependency on a lexer implementation
