+++ yql/essentials/sql/v1/complete/text/ut/ya.make
@@ -1,0 +1,7 @@
+UNITTEST_FOR(yql/essentials/sql/v1/complete/text)
+
+SRCS(
+    word_ut.cpp
+)
+
+END()
