+++ yql/essentials/sql/v1/complete/text/word.cpp
@@ -1,4 +1,4 @@
-#include "string_util.h"
+#include "word.h"
 
 #include <util/generic/strbuf.h>
