+++ yql/essentials/sql/v1/complete/text/word_ut.cpp
@@ -1,10 +1,10 @@
-#include "string_util.h"
+#include "word.h"
 
 #include <library/cpp/testing/unittest/registar.h>
 
 using namespace NSQLComplete;
 
-Y_UNIT_TEST_SUITE(StringUtilTest) {
+Y_UNIT_TEST_SUITE(WordTest) {
     Y_UNIT_TEST(Blank) {
         UNIT_ASSERT_VALUES_EQUAL(LastWord(""), "");
         UNIT_ASSERT_VALUES_EQUAL(LastWord(" "), "");
