+++ yql/essentials/sql/v1/complete/text/ya.make
@@ -1,0 +1,11 @@
+LIBRARY()
+
+SRCS(
+    word.cpp
+)
+
+END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
