+++ yql/essentials/sql/v1/complete/ut/ya.make
@@ -2,7 +2,6 @@ UNITTEST_FOR(yql/essentials/sql/v1/complete)
 
 SRCS(
     sql_complete_ut.cpp
-    string_util_ut.cpp
 )
 
 PEERDIR(
