+++ yql/essentials/sql/v1/complete/ya.make
@@ -5,12 +5,9 @@ SRCS(
     sql_complete.cpp
     sql_context.cpp
     sql_syntax.cpp
-    string_util.cpp
 )
 
 PEERDIR(
-    contrib/libs/antlr4_cpp_runtime
-    contrib/libs/antlr4-c3
     yql/essentials/core/issue
     yql/essentials/sql/settings
     yql/essentials/sql/v1/format
@@ -23,8 +20,10 @@ PEERDIR(
     yql/essentials/parser/antlr_ast/gen/v1_ansi_antlr4
     yql/essentials/parser/antlr_ast/gen/v1_antlr4
 
+    yql/essentials/sql/v1/complete/antlr4
     yql/essentials/sql/v1/complete/name
     yql/essentials/sql/v1/complete/name/static
+    yql/essentials/sql/v1/complete/text
 )
 
 END()
